requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
simplejson==3.19.1
soupsieve==2.4.1
urllib3==2.0.3
